fake evo

+++ ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: ibayandu <ibayandu@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/08 09:18:50 by ibayandu          #+#    #+#             */
-/*   Updated: 2024/10/08 12:13:11 by ibayandu         ###   ########.fr       */
+/*   Updated: 2024/10/12 10:19:26 by ibayandu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, size_t n)
 {
-	ft_memset((void *)s, 0, (unsigned int)n);
+	ft_memset((void *)s, 0, (size_t)n);
 }
