+++ ft_memchr.c
@@ -6,13 +6,15 @@
 /*   By: ibayandu <ibayandu@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/10 18:44:14 by ibayandu          #+#    #+#             */
-/*   Updated: 2024/10/10 18:56:31 by ibayandu         ###   ########.fr       */
+/*   Updated: 2024/10/12 10:20:07 by ibayandu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memchr(const void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memchr(const void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
