+++ ft_memcmp.c
@@ -6,13 +6,15 @@
 /*   By: ibayandu <ibayandu@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/10 18:57:24 by ibayandu          #+#    #+#             */
-/*   Updated: 2024/10/10 19:00:12 by ibayandu         ###   ########.fr       */
+/*   Updated: 2024/10/12 10:20:34 by ibayandu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
