+++ ft_memmove.c
@@ -6,11 +6,13 @@
 /*   By: ibayandu <ibayandu@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/08 09:30:33 by ibayandu          #+#    #+#             */
-/*   Updated: 2024/10/08 13:49:05 by ibayandu         ###   ########.fr       */
+/*   Updated: 2024/10/12 10:21:14 by ibayandu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *dst, const void *src, unsigned int len)
+#include "libft.h"
+
+void	*ft_memmove(void *dst, const void *src, size_t len)
 {
 	char		*dest;
 	char const	*source;
