+++ ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: ibayandu <ibayandu@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/07 17:56:15 by ibayandu          #+#    #+#             */
-/*   Updated: 2024/10/09 22:33:05 by ibayandu         ###   ########.fr       */
+/*   Updated: 2024/10/14 11:42:22 by ibayandu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *dest, int c, unsigned int len)
+#include "libft.h"
+
+void	*ft_memset(void *dest, int c, size_t len)
 {
 	unsigned char	*ptr;
