+++ ft_strlcpy.c
@@ -6,15 +6,15 @@
 /*   By: ibayandu <ibayandu@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/08 11:20:25 by ibayandu          #+#    #+#             */
-/*   Updated: 2024/10/08 13:34:53 by ibayandu         ###   ########.fr       */
+/*   Updated: 2024/10/12 10:22:25 by ibayandu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcpy(char *dst, const char *src, unsigned int dstsize)
+unsigned int	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (dstsize)
