+++ ft_strnstr.c
@@ -6,16 +6,16 @@
 /*   By: ibayandu <ibayandu@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/10 19:02:38 by ibayandu          #+#    #+#             */
-/*   Updated: 2024/10/10 22:38:18 by ibayandu         ###   ########.fr       */
+/*   Updated: 2024/10/12 10:23:01 by ibayandu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strnstr(const char *haystack, const char *needle, unsigned int len)
+char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	unsigned int	i;
-	unsigned int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	if (!haystack && !len)
